backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-socketio==5.10.0
python-dotenv==1.0.0
aiohttp==3.9.1
